+++ packages/webblocks-cms/resources/lang/de/blocks.php
@@ -6,4 +6,13 @@ return [
     'placeholder' => 'Diese Website durchsuchen',
     'submit' => 'Suchen',
   ],
+  'contact_form' => [
+    'submit' => 'Senden',
+    'review' => 'Bitte prüfen Sie das Formular',
+    'name' => 'Ihr Name',
+    'email' => 'Ihre E-Mail',
+    'subject' => 'Betreff',
+    'message' => 'Ihre Nachricht',
+    'storage' => 'Ihre Nachricht wird gespeichert, anschließend wird die E-Mail-Benachrichtigung versucht.',
+  ],
 ];